fix: specify destination address in sendto when sending on raw socket
It is necessary to specify the destination address of a packet to the
sendto syscall, even when sending on an IPPROTO_RAW socket, even though
the information is already present in the packet header.
If the destination address for the sendto call is zero, we noticed that
the packets arrived at the local loopback interface, instead of being
send to the correct addresses.