feat: add registration process with beekeeper
This enables the bee to register itself with the beekeeper.
If not registered yet, it asks the user to input a registration token
which it'll use to contact the beekeeper API. Afterward, the
authentication token and the bee's id is received and stored in a local
file to receive the configuration when the bee is restarted.

This also adds the beekeeper's API spec and generates the endpoint API
part of it automatically using oapi-codegen and go:generate.
This is useful since we don't have to write the API client ourselves and
ensure to be compliant with the API spec.

Moreover, after registering with the API, the API client of the bee is
configured to send the received authentication token as a bearer token
in each subsequent request.