docs: make requirement of libpcap-dev more prominent
This makes the requirement of libpcap-dev more prominent by including a
command for apt install. Since users likely skip the textual sections and
only copy&paste the commands, this should ensure that they actually install
libpcap-dev. Since this only works on Debian, this should also stop them and
make them read for a moment if they are on another system.