fix: make bee work immediatly without having to restart
Previously, we would create a wrong security provider when registering
the bee, which lead to a failure. As a result, the bee wouldn't be able
to start up properly.

We now correctly configure the security provider (we have done so before already,
but when registering attached a second one for a Bearer token which the backend
didn't like).