feat: remove requirement for CAP_SYS_ADMIN
By using netns.None() as the argument to the NewHandleAt,
we ensure that we get a handle to the current namespace.
While this has worked using the netnsFd from netns.Get
previously as well, this apparently required CAP_SYS_ADMIN
privileges (even though we targeted the same namespace we are in).

To require as few privileges as possible and therefore make running
in a Docker container easier (especially since CAP_SYS_ADMIN really
allows our container to do A LOT), this has been changed.