feat: add automatic install and uninstall of nftables
This now automatically installs the required nftables on startup and also
removes them on stop to cleanup properly.

Since we currently do not detect whether legacy iptables are present, users
can disable the automatic nftables generation to modify the firewall themselfs.
This is also useful if anything else than nftables is used.

While we could've used github.com/google/nftables to modify the nftables without
using a subprocess with /usr/sbin/nft, the library seems rather work in progress
right now. Moreover, we also tested an implementation using that library and the
code was way less readable than the current nftables configuration. Since we
shouldn't need to modify the nftables too much, the time (and memory) penalty of
a subprocess shouldn't hurt too much for now.